1.support base restful 2.security function(Resource Server) not test

+++ resource/src/main/java/com/ray/service/resource/controller/ResourceController.java
@@ -3,10 +3,7 @@ package com.ray.service.resource.controller;
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONObject;
 import com.ray.service.resource.clients.RemoteAccountFeignClient;
-import com.ray.service.resource.model.Account;
-import com.ray.service.resource.model.FileInfo;
-import com.ray.service.resource.model.Log;
-import com.ray.service.resource.model.Reply;
+import com.ray.service.resource.model.*;
 import com.ray.service.resource.service.AccountService;
 import com.ray.service.resource.service.ResourceService;
 import com.ray.service.resource.util.ErrInfo;
@@ -46,13 +43,25 @@ public class ResourceController {
         return accountService.loadUserByUsername(principal.getName());
     }
 
+    private List<Group> getAllGroupBelongOpAccount() {
+        String info = accountService.getGroupList();
+        if (info.equals("Invalid resources.")) {
+            return null;
+        }
+        List<Group> groupList = JSON.parseArray(info, Group.class);
+        if (groupList == null || groupList.isEmpty()) {
+            return null;
+        }
+        return groupList;
+    }
+
     @PostMapping("new")
     public String createNewLog(Principal principal,
                                @RequestParam(name = "title", required = true) String title,
                                @RequestParam(name = "type", required = true) int type,
                                @RequestParam(name = "gid", required = false, defaultValue = "0") int gid) {
         Account account = getOpAccount(principal);
-        if(account == null){
+        if (account == null) {
             return ErrInfo.assembleJson(ErrInfo.ErrType.NULLOBJ, ErrInfo.CODE_CREATE_NEW, "Invalid request parameters.");
         }
         Log log = new Log();
@@ -72,11 +81,11 @@ public class ResourceController {
                             @RequestParam(name = "type", required = false, defaultValue = "0") int type,
                             @RequestParam(name = "gid", required = false, defaultValue = "0") int gid) {
         Account account = getOpAccount(principal);
-        if(account == null){
+        if (account == null) {
             return ErrInfo.assembleJson(ErrInfo.ErrType.NULLOBJ, ErrInfo.CODE_UPDATE_RES, "Invalid request parameters.");
         }
         Log log = resourceService.getLogDetial(id);
-        if(log.getUid() != account.getUid()){
+        if (log.getUid() != account.getUid()) {
             return ErrInfo.assembleJson(ErrInfo.ErrType.INVALID, ErrInfo.CODE_UPDATE_RES, "You don't have permission to access.");
         }
         if (!title.isEmpty()) {
@@ -96,11 +105,11 @@ public class ResourceController {
     public String deleteLog(Principal principal,
                             @RequestParam(name = "id", required = true) int id) {
         Account account = getOpAccount(principal);
-        if(account == null){
+        if (account == null) {
             return ErrInfo.assembleJson(ErrInfo.ErrType.NULLOBJ, ErrInfo.CODE_DELETE_RES, "Invalid request parameters.");
         }
         Log log = resourceService.getLogDetial(id);
-        if(log.getUid() != account.getUid()){
+        if (log.getUid() != account.getUid()) {
             return ErrInfo.assembleJson(ErrInfo.ErrType.INVALID, ErrInfo.CODE_DELETE_RES, "You don't have permission to access.");
         }
         resourceService.deleteLog(log);
@@ -116,11 +125,11 @@ public class ResourceController {
                                   @RequestParam(name = "checksum", required = true) String checksum) {
         String result = "";
         Account account = getOpAccount(principal);
-        if(account == null){
+        if (account == null) {
             return ErrInfo.assembleJson(ErrInfo.ErrType.NULLOBJ, ErrInfo.CODE_UPLOAD_RES, "Invalid request parameters.");
         }
         Log log = resourceService.getLogDetial(id);
-        if(log.getUid() != account.getUid()){
+        if (log.getUid() != account.getUid()) {
             return ErrInfo.assembleJson(ErrInfo.ErrType.INVALID, ErrInfo.CODE_UPLOAD_RES, "You don't have permission to access.");
         }
         List<MultipartFile> files = ((MultipartHttpServletRequest) request).getFiles(uploadFileLabel);
@@ -130,7 +139,7 @@ public class ResourceController {
             return ErrInfo.assembleJson(ErrInfo.ErrType.NULLOBJ, ErrInfo.CODE_UPLOAD_RES, "There are too many files here.");
         }
         String saveDir = log.getDetial().getRes_url();
-                //NetFile.getDirAbsolutePath(saveDirRoot, String.valueOf(account.getUid()), String.valueOf(id));
+        //NetFile.getDirAbsolutePath(saveDirRoot, String.valueOf(account.getUid()), String.valueOf(id));
         try {
             int err = NetFile.saveSplitFile(saveDir, filename, index, total, checksum, chunkSize * 1024 * 1024, files.get(0));
             if (err != 0) {
@@ -147,11 +156,11 @@ public class ResourceController {
                              @RequestParam(name = "id", required = true) int id,
                              @RequestParam(name = "files", required = false, defaultValue = "all") String filesname) {
         Account account = getOpAccount(principal);
-        if(account == null){
+        if (account == null) {
             return ErrInfo.assembleJson(ErrInfo.ErrType.NULLOBJ, ErrInfo.CODE_DELETE_RES_FILES, "Invalid request parameters.");
         }
         Log log = resourceService.getLogDetial(id);
-        if(log.getUid() != account.getUid()){
+        if (log.getUid() != account.getUid()) {
             return ErrInfo.assembleJson(ErrInfo.ErrType.INVALID, ErrInfo.CODE_DELETE_RES_FILES, "You don't have permission to access.");
         }
         if (log == null) {
@@ -166,10 +175,10 @@ public class ResourceController {
                                  @RequestParam(name = "id", required = true) int id) {
         Account account = getOpAccount(principal);
         Log log = resourceService.getLogDetial(id);
-        if(account == null || log == null){
+        if (account == null || log == null) {
             return ErrInfo.assembleJson(ErrInfo.ErrType.NULLOBJ, ErrInfo.CODE_GET_RES_FILES, "Invalid request parameters.");
         }
-        if(log.getUid() != account.getUid()){
+        if (log.getUid() != account.getUid()) {
             return ErrInfo.assembleJson(ErrInfo.ErrType.INVALID, ErrInfo.CODE_GET_RES_FILES, "You don't have permission to access.");
         }
         List<FileInfo> fileInfoList;
@@ -185,53 +194,76 @@ public class ResourceController {
     public String createNewReply(Principal principal,
                                  @RequestParam(name = "id", required = true) int id,
                                  @RequestParam(name = "msg", required = true) String msg) {
+        Account account = getOpAccount(principal);
+        Log log = resourceService.getLogDetial(id);
+        if (account == null || log == null) {
+            return ErrInfo.assembleJson(ErrInfo.ErrType.NULLOBJ, ErrInfo.CODE_CREATE_REPLY, "Invalid request parameters.");
+        }
         Reply reply = new Reply();
         reply.setDate(SimpleDate.getUTCTime());
         reply.setDesc(msg);
         reply.setLid(id);
+        reply.setUid(account.getUid());
         resourceService.createReply(reply);
         return JSONObject.toJSONString(reply);
     }
 
     @DeleteMapping("reply")
     public String deleteReply(Principal principal, int lid, int rid) {
-        boolean found = false;
+        Account account = getOpAccount(principal);
         Log log = resourceService.getLogDetial(lid);
-        if (log == null) {
-            return ErrInfo.assembleJson(ErrInfo.ErrType.NULLOBJ, ErrInfo.CODE_GET_LOGFILE, "there are no logfile existes.");
+        if (account == null || log == null) {
+            return ErrInfo.assembleJson(ErrInfo.ErrType.NULLOBJ, ErrInfo.CODE_DELETE_REPLY, "Invalid request parameters.");
         }
-        for(Reply reply:log.getReplyList()){
-            if(reply.getId() == rid){
-                found = true;
-                resourceService.deleteReply(reply);
-                break;
+        for (Reply reply : log.getReplyList()) {
+            if (reply.getId() != rid) {
+                continue;
             }
-        }
-        if(!found) {
-            return ErrInfo.assembleJson(ErrInfo.ErrType.NULLOBJ, ErrInfo.CODE_GET_LOGFILE, "there are no logfile existes.");
+            if (reply.getUid() == account.getUid() || log.getUid() == account.getUid()) {
+                resourceService.deleteReply(reply);
+            }
+            break;
         }
         return "";
     }
 
     @GetMapping("replies")
-    public String getAllReply(@RequestParam(name = "id", required = true)int id) {
+    public String getAllReply(@RequestParam(name = "id", required = true) int id) {
         List<Reply> replies = resourceService.getReplies(id);
         return JSONObject.toJSONString(replies);
     }
 
     @GetMapping("list")
     public String getResource(Principal principal,
-                              @RequestParam(name="type", required = false, defaultValue = "0")int type,
-                              @RequestParam(name="gid", required = false, defaultValue = "0")int gid) {
+                              @RequestParam(name = "type", required = false, defaultValue = "0") int type,
+                              @RequestParam(name = "gid", required = false, defaultValue = "0") int gid) {
+        Account account = getOpAccount(principal);
         List<Log> logs;
-        if(type != 0 && gid != 0){
-            logs = resourceService.getGroupLogsByType(type, gid);
-        }else if(type != 0){
+        if (account == null && type != 0) {
             logs = resourceService.getPublicLogsByType(type);
-        }else{
-//            logs = resourceService.getLogsByUid();
+        } else if (account == null) {
+            return ErrInfo.assembleJson(ErrInfo.ErrType.INVALID, ErrInfo.CODE_GET_RES, "You don't have permission to access.");
+        } else if (type != 0 && isBelongGid(gid)) {
+            logs = resourceService.getGroupLogsByType(type, gid);
+        } else if (type != 0) {
+            logs = resourceService.getPublicLogsByType(type);
+        } else {
+            logs = resourceService.getLogsByUid(account.getUid());
         }
         return JSONObject.toJSONString(logs);
+    }
+
+    private boolean isBelongGid(int gid){
+        List<Group> groupList = getAllGroupBelongOpAccount();
+        if(groupList == null || gid == 0){
+            return false;
+        }
+        for(Group group:groupList){
+            if(group.getGid() == gid){
+                return true;
+            }
+        }
+        return false;
     }
 
 }
